Fix commeter name change
