+++ app/(website)/account/[tab]/EmailPreferences.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import {useState, useEffect, useReducer} from "react";
+import {useState, useEffect} from "react";
 
-import styles from "@/styles/pages/account.module.scss";
+import LoadingOverlay from "./LoadingOverlay";
 
 const EmailPreferences = ({user}) => {
   const [selectedNewsLetters, setSelectedNewsLetters] = useState<String[]>([]);
@@ -143,26 +143,5 @@ const EmailPreferences = ({user}) => {
     </>
   );
 };
-
-const LoadingOverlay = ({message}: {message: string}) => (
-  <div
-    style={{
-      position: "absolute",
-      top: "50%",
-      left: "50%",
-      transform: "translate(-50%, -50%)",
-      zIndex: "100",
-      backgroundColor: "#FFF",
-      color: "#000",
-      border: "1px solid #000",
-      padding: "10px",
-      display: "flex",
-      justifyContent: "center",
-      alignItems: "center",
-      maxWidth: "300px",
-    }}>
-    {message}
-  </div>
-);
 
 export default EmailPreferences;
