+++ app/(website)/account/[tab]/page.tsx
@@ -3,6 +3,7 @@ import {
   getSession,
   getUserDetails,
   getSubscription,
+  getProfile,
 } from "@/app/(website)/supabase-server";
 import type {Database} from "@/types/supabase";
 import {createServerActionClient} from "@supabase/auth-helpers-nextjs";
@@ -43,6 +44,7 @@ export default async function Account({params}) {
     getUserDetails(),
     getSubscription(),
   ]);
+  const profile = await getProfile(userDetails?.id!);
   const {tab} = params;
 
   const globalFields = await getGlobalFields();
@@ -82,7 +84,7 @@ export default async function Account({params}) {
 
     if (user) {
       const {error} = await supabase
-        .from("users")
+        .from("profiles")
         .update({comments_display_name: newDisplayName})
         .eq("id", user.id);
 
@@ -144,6 +146,7 @@ export default async function Account({params}) {
         userDetails={userDetails}
         subscription={subscription}
         session={session}
+        profile={profile}
         updateName={updateName}
         updateEmail={updateEmail}
         updateCommentsNotifications={updateCommentsNotifications}
